+++ tests/Browser/Pages/HomePage.php
@@ -35,7 +35,39 @@ class HomePage extends Page
     public function elements()
     {
         return [
-            '@element' => '#selector',
+            '@newBorrowingButton' => '#new-borrowing-button',
+            '@endBorrowingButton' => '#end-borrowing-button',
+            '@borrowingsHistoryButton' => '#borrowings-history-button',
+            '@viewInventoryButton' => '#view-inventory-button',
+            '@editInventoryButton' => '#edit-inventory-button',
+            '@editUsersButton' => '#edit-users-button',
+            '@githubLink' => '#github-link'
         ];
+    }
+
+    public function navigateTo(Browser $browser, $page) {
+        switch ($page) {
+            case 'new-borrowing':
+                $browser->click('@newBorrowingButton');
+                break;
+            case 'end-borrowing':
+                $browser->click('@endBorrowingButton');
+                break;
+            case 'borrowings-history':
+                $browser->click('@borrowingsHistoryButton');
+                break;
+            case 'view-inventory':
+                $browser->click('@viewInventoryButton');
+                break;
+            case 'edit-inventory':
+                $browser->click('@editInventoryButton');
+                break;
+            case 'edit-users':
+                $browser->click('@editUsersButton');
+                break;
+            case 'github':
+                $browser->click('@githubLink');
+                break;
+        }
     }
 }