+++ changelog.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * Ids to buttons
 * Testing
     * Dusk setup.
-    * Integration tests for pages visits.
+    * Unit tests for pages visits.
+    * Unit tests for pages accesses from home page.
 
 ## [0.13.0] - 2018-10-07